+++ KGTorrent/kgtorrent.py
@@ -4,15 +4,11 @@ import argparse
 import sys
 from pathlib import Path
 
-from sqlalchemy_utils import database_exists, \
-    create_database, \
-    drop_database
-
-from KGTorrent import config, \
-    build_db_schema, \
-    populate_db, \
-    downloader
-from KGTorrent.db_connection_handler import DbConnectionHandler
+import KGTorrent.config as config
+from KGTorrent.data_loader import DataLoader
+from KGTorrent.db_communication_handler import DbCommunicationHandler
+from KGTorrent.downloader import Downloader
+from KGTorrent.mk_preprocessor import MkPreprocessor
 
 
 def main():
@@ -58,16 +54,24 @@ def main():
 
     command = args.command
 
+    print("************************")
+    print("*** KGTORRENT STARTED***")
+    print("************************")
+
     # Create db engine
-    db_connection_handler = DbConnectionHandler()
-    db_engine = db_connection_handler.create_sqlalchemy_engine()
-    print("Connected to the database")
+    print(f"## Connecting to {config.db_name} db on port {config.db_port} as user {config.db_username}")
+    db_engine = DbCommunicationHandler(config.db_username,
+                                       config.db_password,
+                                       config.db_host,
+                                       config.db_port,
+                                       config.db_name)
+    print("Connection established.")
 
     # CHECK USER VARIABLES
     proceed = None
 
     # Check db emptiness
-    if database_exists(db_engine.url):
+    if db_engine.db_exists():
         if command == 'init':
             print(f'Database {config.db_name} already exists. ', file=sys.stderr)
             print(f'Please, provide a name that is not already in use for the KGTorrent database.',
@@ -91,50 +95,52 @@ def main():
         print('Please, provide the path to an empty folder to store downloaded notebooks.', file=sys.stderr)
         proceed = False
 
-    #KGTorrent process
+    # KGTorrent process
     if proceed:
 
-        print('Initializing the database...')
-        # Initialize the db
-        if database_exists(db_engine.url):
-            drop_database(db_engine.url)
-        create_database(db_engine.url,'utf8mb4')
+        print("********************")
+        print("*** LOADING DATA ***")
+        print("********************")
+        dl = DataLoader(config.constraints_file_path, config.meta_kaggle_path)
 
-        print('Building table schemas...')
-        # Build the database schema
-        db_schema = build_db_schema.DbSchema(
-            sqlalchemy_engine=db_engine
-        )
-        print("Database schema built")
+        print("**********************************")
+        print("** TABLES PREPROCESSING STARTED **")
+        print("**********************************")
+        mk = MkPreprocessor(dl.get_tables_dict(), dl.get_constraints_df())
+        processed_dict, stats = mk.preprocess_mk()
 
-        # Populate the database with notebook metadata
-        mk_preprocessor = populate_db.MetaKagglePreprocessor(
-            config.constraints_file_path,
-            meta_kaggle_path=config.meta_kaggle_path,
-            sqlalchemy_engine=db_engine)
-        populate_db.populate_db(
-            mk_preprocessor)
+        print("*************")
+        print("*** STATS ***")
+        print("*************\n")
+        print(stats)
 
-        print('Applying foreign key constraints')
-        populate_db.set_foreign_keys(
-            sqlalchemy_engine=db_engine,
-            constraints_file_path=config.constraints_file_path)
+        print("Initializing DB...")
+        db_engine.create_new_db(drop_if_exists=True)
 
-        print("Database populated with MetaKaggle data")
+        print("***************************")
+        print("** DB POPULATION STARTED **")
+        print("***************************")
+        db_engine.write_tables(processed_dict)
+
+        print("** APPLICATION OF CONSTRAINTS **")
+        db_engine.set_foreign_keys(dl.get_constraints_df())
+
+        print("** QUERING KERNELS TO DOWNLOAD **")
+        kernels_ids = db_engine.get_kernels_identifiers(config.download_conf['languages'])
 
         # Download the notebooks and update the db with their local path
         # To get a specific subset of notebooks, query the database by using
         # the db_schema object as needed.
-        print("Notebooks download started")
-        downloader.Downloader(
-            sqlalchemy_engine=db_engine,
-            download_config=config.download_conf,
-            strategy=args.strategy
-        )
-        print("Notebooks download completed")
+        print("*******************************")
+        print("** NOTEBOOK DOWNLOAD STARTED **")
+        print("*******************************")
+        downloader = Downloader(kernels_ids.head(), config.nb_archive_path)
+        print(f'# Selected strategy. {args.strategy}')
+        downloader.download_kernels(strategy=args.strategy)
+        print('## Download finished.')
 
     time.sleep(0.2)
-    print('KGTorrent end')
+    print('## KGTorrent end')
 
 
 if __name__ == '__main__':
